+++ example/a.ts
@@ -1,5 +1,20 @@
 import { Elysia, t } from '../src'
 
 const app = new Elysia()
-	.group('/v1', (app) => app.decorate('a', 'b').get('/', ({ a }) => a))
+	.get('/', () => 'Mutsuki need correction 💢💢💢', {
+		afterHandle() {
+			return 'Mutsuki need correction 💢💢💢'
+		},
+		schema: {
+			response: t.String()
+		}
+	})
+	.get('/invalid', () => 1 as any, {
+		afterHandle() {
+			return 1 as any
+		},
+		schema: {
+			response: t.String()
+		}
+	})
 	.listen(8080)